// core/types.ts
export enum EducationalStage {
  PRE_SCHOOL = 'PRE_SCHOOL',
  PRIMARY = 'PRIMARY',
  MIDDLE = 'MIDDLE',
  HIGH = 'HIGH',
}

export enum Subject {
  // Primary
  Arabic = 'اللغة العربية',
  French = 'اللغة الفرنسية',
  Math = 'الرياضيات',
  IslamicEducation = 'التربية الإسلامية',
  SocialStudies = 'الاجتماعيات',

  // Middle & High
  English = 'اللغة الإنجليزية',
  Physics = 'الفيزياء والكيمياء',
  NaturalSciences = 'علوم الحياة والأرض',
  ComputerScience = 'المعلوميات',
  Philosophy = 'الفلسفة',
}
// ...
interface BaseContent {
  id: number;
  stage: EducationalStage;
  level: string;
  class: string;
  subject: Subject;
}
// ...
export interface Note extends BaseContent {
  studentIds: string[];
  observation: string;
  image?: string;
  pdf?: { name: string; url: string };
  externalLink?: string;
  date: Date;
  status: 'pending' | 'approved';
  type?: 'ai_report';
}

export interface Absence {
    id: number;
    studentId: string;
    date: Date;
    stage: EducationalStage;
    level: string;
    class: string;
    subject: Subject;
}

export interface ExamProgram extends BaseContent {
    date: Date;
    image?: string; // base64 string
    pdf?: { name: string; url: string }; // name and blob url
}

export interface Grade {
    subSubject: string;
    semester: 1 | 2;
    assignment: 1 | 2;
    score: number | null;
}

export interface Student {
    id: string;
    guardianCode: string;
    name: string;
    stage: EducationalStage;
    level: string;
    class: string;
    grades: {
        [key in Subject]?: Grade[];
    };
}